deploy/autofigure-sidecar/requirements-upstream.txt
# AutoFigure - Core Dependencies
# Install with: pip install -r requirements.txt

# Core LLM providers
openai>=1.0.0
anthropic>=0.18.0
google-genai>=0.3.0

# Image processing
Pillow>=10.0.0
cairosvg>=2.7.0

# PDF processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# HTTP requests
requests>=2.31.0

# Browser automation (for mxGraph XML to PNG conversion)
# After installing, run: playwright install chromium
playwright>=1.40.0
